whatsapp

+++ DMS/Appointment.aspx.cs
@@ -863,22 +863,31 @@ namespace DMS
             Page.Validate();
             if (Page.IsValid == true)
             {
-                SqlConnection con = new SqlConnection(strCon);
-                try
+                string confirmValue = Request.Form["confirm_value"];
+
+                if (confirmValue == "Yes")
                 {
-                    con.Open();
-                    SqlCommand cmd = new SqlCommand("DELETE FROM Appointment WHERE appointmentID = @appointmentID", con);
-                    cmd.Parameters.AddWithValue("@appointmentID", txtDeleteAppointID.Text);
-                    cmd.ExecuteNonQuery();
-                    Response.Write("<script type=\"text/javascript\">alert('Appointment details have been successfully deleted.');location.href='Appointment.aspx'</script>");
+                    SqlConnection con = new SqlConnection(strCon);
+                    try
+                    {
+                        con.Open();
+                        SqlCommand cmd = new SqlCommand("DELETE FROM Appointment WHERE appointmentID = @appointmentID", con);
+                        cmd.Parameters.AddWithValue("@appointmentID", txtDeleteAppointID.Text);
+                        cmd.ExecuteNonQuery();
+                        Response.Write("<script type=\"text/javascript\">alert('Appointment details have been successfully deleted.');location.href='Appointment.aspx'</script>");
+                    }
+                    catch (SqlException ex)
+                    {
+                        AlertMessage(ex.Message);
+                    }
+                    finally
+                    {
+                        con.Close();
+                    }
                 }
-                catch (SqlException ex)
+                else
                 {
-                    AlertMessage(ex.Message);
-                }
-                finally
-                {
-                    con.Close();
+                    Response.Write("<script type=\"text/javascript\">alert('Appointment Details have not been successfully deleted.')</script>");
                 }
             }
         }
